+++ term/cmd.go
@@ -1,0 +1,6 @@
+package term
+
+var CommandDescriptions = map[string]string{
+	"docs": "use 'fizz docs [library name]' to print out a full list of functons declared in that library. example: 'fizz docs io'",
+	"help": "prints a longer description for a given command. example 'fizz help docs'",
+}